uloha 6

+++ src/main/java/sk/ukf/mvc/service/EmployeeServiceImpl.java
@@ -1,8 +1,11 @@
 package sk.ukf.mvc.service;
 
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 import sk.ukf.mvc.dao.EmployeeRepository;
 import sk.ukf.mvc.entity.Employee;
+import sk.ukf.mvc.exception.EmailAlreadyExistsException;
+import sk.ukf.mvc.exception.ObjectNotFoundException;
 
 import java.util.List;
 
@@ -21,31 +24,32 @@ public class EmployeeServiceImpl implements EmployeeService {
 
     @Override
     public Employee findById(int id) {
-//        return employeeRepository.findById(id).orElseThrow(() -> new ObjectNotFoundException("Zamestnanec", id));
-        return employeeRepository.findById(id).orElse(null);
+        return employeeRepository.findById(id).orElseThrow(() -> new ObjectNotFoundException("Zamestnanec", id));
     }
 
+    @Transactional
     @Override
     public Employee save(Employee employee) {
-//        if (employee.getId() == 0) {
-//            if (employeeRepository.existsByEmail(employee.getEmail())) {
-//                throw new EmailAlreadyExistsException(employee.getEmail());
-//            }
-//        } else {
-//            Employee existingWithEmail = employeeRepository.findByEmail(employee.getEmail()).orElse(null);
-//            if (existingWithEmail != null && existingWithEmail.getId() != employee.getId()) {
-//                throw new EmailAlreadyExistsException(employee.getEmail());
-//            }
-//        }
+        if (employee.getId() == 0) {
+            if (employeeRepository.existsByEmail(employee.getEmail())) {
+                throw new EmailAlreadyExistsException(employee.getEmail());
+            }
+        } else {
+            Employee existingWithEmail = employeeRepository.findByEmail(employee.getEmail()).orElse(null);
+            if (existingWithEmail != null && existingWithEmail.getId() != employee.getId()) {
+                throw new EmailAlreadyExistsException(employee.getEmail());
+            }
+        }
 
         return employeeRepository.save(employee);
     }
 
+    @Transactional
     @Override
     public void deleteById(int id) {
-//        if (!employeeRepository.existsById(id)) {
-//            throw new ObjectNotFoundException("Zamestnanec", id);
-//        }
+        if (!employeeRepository.existsById(id)) {
+            throw new ObjectNotFoundException("Zamestnanec", id);
+        }
         employeeRepository.deleteById(id);
     }
 }
